requirements.txt
Django==3.2.9
django-phonenumber-field==6.0.0
django-taggit==2.0.0
phonenumbers==8.12.40
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
requests==2.27.1
slugify==0.0.1
virtualenv==20.13.0


xhtml2pdf~=0.2.7
pdfkit~=1.0.0